Adds Springboot to project
