+++ src/main/java/gildedroseinventory/ConjuredItem.java
@@ -1,20 +1,23 @@
-public class AgedBrieItem implements Itemable{
+package gildedroseinventory;
+
+public class ConjuredItem implements Itemable{
     public final String name;
     public int sellIn;
     public int quality;
 
-    public AgedBrieItem(String name, int sellIn, int quality) {
+    public ConjuredItem(String name, int sellIn, int quality) {
         this.name = name;
         this.sellIn = sellIn;
         this.quality = quality;
     }
+
     @Override
     public void update() {
         this.sellIn -= 1;
         if (this.sellIn >= 0) {
-            this.quality = Math.min(50, this.quality + 1);
+            this.quality = Math.max(0, this.quality -= 2);
         } else {
-            this.quality = Math.min(50, this.quality + 2);
+            this.quality = Math.max(0, this.quality -= 4);
         }
     }
 }
