+++ src/main/java/gildedroseinventory/Itemable.java
@@ -1,3 +1,5 @@
+package gildedroseinventory;
+
 public interface Itemable {
     public void update();
 }
