+++ src/test/java/AgedBrieItemTest.java
@@ -1,3 +1,4 @@
+import gildedroseinventory.AgedBrieItem;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
