+++ src/test/java/BackstagePassesItemTest.java
@@ -1,3 +1,4 @@
+import gildedroseinventory.BackstagePassesItem;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
