+++ src/test/java/ConjuredItemTest.java
@@ -1,3 +1,4 @@
+import gildedroseinventory.ConjuredItem;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
