+++ src/test/java/GildedRoseTest.java
@@ -1,3 +1,5 @@
+import gildedroseinventory.GildedRose;
+import gildedroseinventory.Itemable;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
