+++ src/test/java/NormalItemTest.java
@@ -1,3 +1,4 @@
+import gildedroseinventory.NormalItem;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
