+++ src/test/java/SulfurasItemTest.java
@@ -1,3 +1,4 @@
+import gildedroseinventory.SulfurasItem;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
